Refactor, removing unnecessary file
